Sequenced loop u16 (#231)
Sending over 65536 packets would previously just saturate the top_index
which caused it to not accept any more packets. This patch shortens the
sequence acceptance by half but always loops around the max u16 value.

This means that at top_index = 0, all sequenced packets 0-32768 are
accepted, while at top_index = 50000, all sequenced packets 50000-65535
and 0-17233 will be accepted.

+++ src/net/virtual_connection.rs
@@ -139,17 +139,16 @@ impl VirtualConnection {
                         );
 
                         if let OrderingGuarantee::Ordered(stream_id) = ordering_guarantee {
-                            let item_identifier = if let Some(item_identifier) =
-                                last_item_identifier
-                            {
-                                item_identifier
-                            } else {
-                                self.ordering_system
-                                    .get_or_create_stream(
-                                        stream_id.unwrap_or(DEFAULT_ORDERING_STREAM),
-                                    )
-                                    .new_item_identifier() as u16
-                            };
+                            let item_identifier =
+                                if let Some(item_identifier) = last_item_identifier {
+                                    item_identifier
+                                } else {
+                                    self.ordering_system
+                                        .get_or_create_stream(
+                                            stream_id.unwrap_or(DEFAULT_ORDERING_STREAM),
+                                        )
+                                        .new_item_identifier()
+                                };
 
                             item_identifier_value = Some(item_identifier);
 
@@ -157,17 +156,16 @@ impl VirtualConnection {
                         };
 
                         if let OrderingGuarantee::Sequenced(stream_id) = ordering_guarantee {
-                            let item_identifier = if let Some(item_identifier) =
-                                last_item_identifier
-                            {
-                                item_identifier
-                            } else {
-                                self.sequencing_system
-                                    .get_or_create_stream(
-                                        stream_id.unwrap_or(DEFAULT_SEQUENCING_STREAM),
-                                    )
-                                    .new_item_identifier() as u16
-                            };
+                            let item_identifier =
+                                if let Some(item_identifier) = last_item_identifier {
+                                    item_identifier
+                                } else {
+                                    self.sequencing_system
+                                        .get_or_create_stream(
+                                            stream_id.unwrap_or(DEFAULT_SEQUENCING_STREAM),
+                                        )
+                                        .new_item_identifier()
+                                };
 
                             item_identifier_value = Some(item_identifier);
